Further refactorings
